organize the different implementations
