+++ .eslintignore
@@ -7,4 +7,4 @@ dist
 !*.*ts
 
 memoize-fs.js
-globbing.js
+# globbing.js
