+++ packages/eslint/src/barry/foo.js
@@ -1,0 +1,1 @@
+module.exports = 123;
